+++ udata/harvest/backends/ckan/__init__.py
@@ -1,0 +1,3 @@
+"""
+CKAN integration for udata
+"""